+++ src/utils/reactUtils.js
@@ -11,7 +11,3 @@ export function useTitle(title) {
         }
     })
 }
-
-export function isDev() {
-    return !process.env.NODE_ENV || process.env.NODE_ENV === 'development';
-}
