+++ src/utils/utils.js
@@ -4,3 +4,7 @@ export function dateToRelativeDate(date) {
     const myMoment = DateTime.fromISO(date);
     return myMoment.setLocale('ca').toRelativeCalendar()
 }
+
+export function isDev() {
+    return !process.env.NODE_ENV || process.env.NODE_ENV === 'development';
+}
